Updated analysis/TODO to remove things that have gotten done Got low level code generation working for some new AST chapel files   (with the same .good test files used with the standard code generator). Fixed bugs in mapping of types from the AST to the low level representation. Added support for "bool" as a 1 bit unsigned numeric type including   support for constant folding in the old parser. Fixed bug in assignment to variables of primitive type.
**** NOTE ****

When we want to overload assignment for other value objects, we
are going to need an assignment function.  If we do it like C++
we can reuse a "copy constructor" which takes a single argument of
the same type.

Currently assignment is always interpreted as assigment to a
primitive numeric type.  To overcome this limitation and permit
assignment to be analyzed, assignment needs to be mapped to an
overloaded function, and that overloaded function needs to be
defined for primitive numeric types to correspond to the builtin
version like (from the old prelude.chpl):

function operator(a:anynum, #"=", b:anynum) { return __operator(a, #"=", b:anynum); }

The __operator represents primitive assignment while other objects can
overload the "operator" function.

This is just a sample syntax, and we don't even have to have make this
public, but it would be easiest for analysis if something like this
existed in the internal prelude.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@804 3a8e244f-b0f2-452b-bcba-4c88e055c3ca