Removed Shannon's .cvsignore which I think she checked in by accident (Shannon, if this isn't true, let me know).
TIP: I find a .cvsignore in the root level of
a project can be very helpful, but also very
user-specific, so I suggest making your own
.cvsignore file which contains .cvsignore,
and then you can add stuff to it, and it'll
never try to check itself in.  Hope that's
true now that a .cvsignore file exists in
this directory.  If not I'll yank it from the
CVS tree manually.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@857 3a8e244f-b0f2-452b-bcba-4c88e055c3ca