Checking in some code I penciled in Friday, but didn't get to check in due to networks going down.
Revamped some symboltable stuff to make the type/level
components of a SymScope to be more appropriate and rely
less on the integer values of the scope types.  Added a
scope after all user code is parsed so that if new
symbols are added after parsing they will be put into this
scope by default.  We may want to put in an assertion that
this scope is empty at codegen time, but I haven't done
so yet.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@873 3a8e244f-b0f2-452b-bcba-4c88e055c3ca