Removed operators &&, ||, !, and % from the compiler, replacing them with "and", "or", "not", and "mod" as per the language spec.  Note mod is not in the language spec.  Updated some test cases to reflect this change.
Eliminated the last two reduce-reduce errors in the parser.  These had
to do with problems parsing index variables in domains.  They are now
parsed as expressions and it is made sure that they are just tuples or
variables.  For example, [i,j in D] and [(i,j) in D] are legal, but
[i+j in D] is not.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@901 3a8e244f-b0f2-452b-bcba-4c88e055c3ca