Refactored much of the code implemented previously for integer configuration variable code generation.
All of the code but one line was moved out of main into
separate files.  Main calls the parseArgs function (in arg.c)
which parses the command-line argument and stores it in
a hash table.  The hash table is made up of configVar structs
with the variable name and value stored as strings.

The setInCommandLine function (in config.c) now simply checks
the table to determine if the variable has been set.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@909 3a8e244f-b0f2-452b-bcba-4c88e055c3ca