Removed the shift-reduce error associated with the else-statement of an if-statement.  Added the optional keyword "then" which can be used between the expression and the then-statement in an "if" statement. The AST is unchanged.  Added two test cases for using if, then, and else.
Removed parsing support for functions with no known type as given by
the syntax "function foo(): { }" in which a type did not follow the
colon.  This removes two problematic shift-reduce errors.

Added two scripts to the test directory, cat_tests and cat_goods,
which print to standard out all of the test programs and their good
outputs.  On each line, the file name and line number are displayed.
These scripts allow for easy grepping of features in the test
directory.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@911 3a8e244f-b0f2-452b-bcba-4c88e055c3ca