Updated parsing of domains so that they can be initialized by sequences but not by constructs enclosed in square brackets.  So
  var D : domain(1) = (1..n);

is legal but

  var D : domain(1) = [1..n];

is not.  This is more inline with the current thinking of Chapel.

Added a stopgap measure to the parser function that creates variables
and constants.  If the variable or constant has a domain type, the
initializer expression is converted into a domain.  This is not the
place to handle this, but this meshes with how domains and sequences
are currently handled.

This checking brings the total number of shift/reduce errors in the
parser down to zero.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@917 3a8e244f-b0f2-452b-bcba-4c88e055c3ca