Added -l and -L flags to the compiler which are fed to the back-end C compiler, allowing additional libraries to be linked into the generated C code.  For example, to run the generated executable using efence, one could do:
	chpl myfile.chpl -lefence

(though on the Cray systems, you would also have to use
the -L flag and LD_LIBRARY_PATH envvar to get efence to
work properly...  Hopefully this module will be touched
up to fix this).

This replaces the previous use of -l which was to specify
logs.  Switched this flag to -d (think of them as "debugging
logs"), and removed the old -d flag (which nobody was using
anyway... and wouldn't you really prefer a log to debug
anyway?  :).  And of course the longer form --log still
works (and I changed the flowanalysis tests to use this
form in case someone reclaims -d later...).

Touched up the arg.cpp file so that if no long-form of a
flag is supplied (-l and -L currently have no long form),
the printout is cleaner.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@939 3a8e244f-b0f2-452b-bcba-4c88e055c3ca