Added initial support for methods.  These don't work yet but are stored.
The main addition in this checkin is a new pass MethodsToFunctions
that should eventually change method calls to function calls, or
something like that.

Added a couple of tests in test/classes/deitz/future that will not yet
be run but show the initial code that I am trying to get to work.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@951 3a8e244f-b0f2-452b-bcba-4c88e055c3ca