Added the ability to assign values to enumerated type symbols.  The values by default start at zero:
	enum adjectives = good, bad, soso;

The values are good is 0, bad is 1, and soso is 2.  These values can
now be overridden as in:

	enum adjectives = good, bad = 2, soso;

In this case, good is 0, bad is 2, and soso is 3.

Added a test case to make sure this works.  Also, moved
test_enum2.chpl out of the no-testing future directory.

Note that the syntax is slightly different than in the spec in which
the initial equal does not belong but the symbols should be wrapped in
curly brackets.  This should be changed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@959 3a8e244f-b0f2-452b-bcba-4c88e055c3ca