Extended the intVal and isComputable functions to work on unary expressions so expressions like -1 are now computable.  Made the decision that the unary operators -, +, and ! on computable integers return computable integers but that the unary operator ~ does not.  Is this a good choice?  My reasoning is for cross-compilation and portability.
Added a couple more tests on enumerated types in which the enumerator
value is negative or not computable.  In the latter case, the compiler
generates a descriptive error message.

Changed the USR_FATAL function so it now prints out a newline at the
end of the error message.  Removed newline symbols from the calls to
this function.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@965 3a8e244f-b0f2-452b-bcba-4c88e055c3ca