Added immature support for records.  In particular, assignment over records is not supported and the compiler needs to be run with the -noanalysis flag when using records.  Methods are not yet supported in records.
Accessing fields from records is supported and is handled using "."
rather than "->" in the generated C code.

Eliminated support for inheritance in classes.  It was only really
implemented in the parser.  It no longer is since the scheme has
changed.  A parent pointer remains in the AST (although this will
probably change eventually too).

Got rid of a print line I accidentally left in the
fieldsToMemberAccesses pass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@999 3a8e244f-b0f2-452b-bcba-4c88e055c3ca