Added a new pass called Fixup that sets back-pointers from expressions to the statements they are in (previously done in SetupStmtLinks) and that sets back-pointers from statements to the symbols they are in (previously done in a limited way within the parser).
Removed two passes SetupStmtLinks and CheckBackLinks which are no
longer necessary.  Though not with this checkin, the Fixup pass will
eventually take a parameter which can be optionally set in order that
the Fixup pass merely check that everything is in order, not put
everything in order.  This support will be added when passes can take
parameters (see one of Brad's future commits).

Changed the function symbol pointer of statements to point to, more
generally, a symbol.  This symbol can be either a function or a class
(since a class has a statement body as well).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1011 3a8e244f-b0f2-452b-bcba-4c88e055c3ca