Added compiler support for simple usage of "with" inside of a class. Variable fields are now copied from a class into another class.  The variables are put in the scope of the new class.  This support should also work with records, though the tests are faltering.
Added a new traversal called DefineSymbols which takes a scope and
adds symbols to it that are defined in the statements being traversed.
This currently only handles VarDefStmts and the variables they define.

Added two new functions to the ILink, preinsert and postinsert.  Only
preinsert is completed and tested.  This inserts a statement LIST into
another statement LIST before a specified statement.  The function
postinsert should do the same thing after the specified statement.

Added a new test case test_class_with1 that successfully tests a
simple usage of the with construct in classes.  Added two new
unsuccessful tests in a future directory that act on records.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1015 3a8e244f-b0f2-452b-bcba-4c88e055c3ca