This checkin makes a few small improvements and prepares to support "with" on records and classes with functions.
The "_this" is now "this" and it is added to the scopes of the
functions in which it is inserted.

The DefineSymbols traversal now inserts function symbols into the new
scope (though this doesn't seem to be found by all subsequent passes).

The fixup pass now assigns non-null parentSymbols to every statement
that points to either a class/record, function, or module depending on
where the statement is located.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1017 3a8e244f-b0f2-452b-bcba-4c88e055c3ca