Add a _this variable to FnSymbol to point to the added "this" field. This also serves as an indicator that the function is a bound function.
For non-primitive value classes declared as variables generate a "new"
primitive to allocate the object.

Complex numbers of various sizes are now registered in IF1 so that
they can be declared as variables.

Removed --noanalysis flag from classes/deitz/future.... this
results in 2 errors, both of which are caused by the way
x.grow() is handled.  I have already sent mail to Steve on this,
and the fix is simple depending on what the language spec says.
I think it is OK to checkin with broken tests if the problems are
known and we can expect to solve them relatively quickly.

This checkin also fails for everything which uses array indexing.
The problem is that the "fixup" pass is not being run on the internal
prelude.  Again, Steve knows about this, but I want to checkpoint
my work, hence the checkin.

Changed the way MemberAcceses are handled.  I now convert
 x.foo(...)

to be foo(x, ...)

We need to rename all member accesses "foo" to be something like "_member_foo"
as agreed in the last meeting to put them in different namespaces.

Fixed a bug in IF1 printing which was mislabelling symbols.

Fixed test cases classes/deitz/future/tesT_method_chpl, test_record1.chpl
  test_method2.chpl by handling instance variables and the new
  "this" parameter.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1023 3a8e244f-b0f2-452b-bcba-4c88e055c3ca