Changed passes so that they are passed the complete list of modules including the two internal modules corresponding to the internal prelude and standard prelude by default.  A pass may then choose to ignore the internal modules in its run() method by checking the "internal" flag that's set on a module symbol if it's internal.
Added a mechanism for traversals so that if they
want to use the default run() method, but want to
skip the internal modules, then they can set their
Traversal::processInternalModules flag to false
at constructor time.

Used this flag for fieldsToMemberAccesses, which
currently doesn't work when called on internal
modules (but Steve will hopefully have it in working
order soon enough).

Also, moved the compilation of the generated C code
into its own pass -- buildBinary -- previously this
was part of the Codegen pass, which seems inappropriate.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1029 3a8e244f-b0f2-452b-bcba-4c88e055c3ca