Fixed a bug in methodsToFunctions in which everything after the method was removed from a class type, not just the function.  Added a new method to ILink called extract which removes a statement from a statement list.
Improved the implementation of copy and copyList for statements.
These procedures can now take an optional scope.  If set, this is used
as the scope for the copied statements, expressions, symbols, and
types.

Implemented a test against codegening with statements.

Renamed ResolveSymbols2 to ResolveSymbols.  Originally this pass was
just going to run after analysis and a pass called ResolveSymbols1 was
going to run before analysis.  Now this single pass can be run
anywhere and it will behave differently if run before or after (in
that it will only use information gleaned from analysis after analysis
has run).  This pass currently only does anything after analysis; it
resolves members in MemberAccesses then.

Added a static field to the RunAnalysis pass that records whether it
has been run or not.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1031 3a8e244f-b0f2-452b-bcba-4c88e055c3ca