Removed the disable flag on fieldsToMemberAccesses so it now runs on the internal_prelude.  This pass, among other things, puts "this" around members in bound functions.
Fixed a bug in which this was put around this (to transform it into
this.this).  This now does not happend, though I think bad things may
still happen to this.a (becoming maybe this.this.a).

Unfortunately, the 7 tests in trivial and flowanalysis are still not
working.  However they seem to be closer in the diff.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1035 3a8e244f-b0f2-452b-bcba-4c88e055c3ca