Added new (broken?) replace functionality for statements.  I'm checking in for synching purposes.  It works fine for the checked in tests.  Next step: Unify replace across the different AST subtypes (statements, expressions, symbols, types, etc.) and make sure they are right.
Changed with resolution to use new statement replace.  This gets rid
of the with-statement (which was previously left).  It works as well
as the other way of doing it.

Fixed a bug in which the scope of "this" added to bound function
parameter lists was incorrectly set.  It is now set to the parameter
scope (where it is defined).  Hopefully this won't be decoupled for
too much longer--See Brad's next checkin.

This checkin resolves seven of the nine failures we've been having.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@1061 3a8e244f-b0f2-452b-bcba-4c88e055c3ca